build(rabbitmq): update amqp libs
Update amqplib and amqp-connection-manager to the latest version.
There are many fixes and improvements.

BREAKING CHANGE: We will no longer emit a disconnect event on an initial connection failure -
instead we now emit connectFailed on each connection failure, and only emit disconnect when we
transition from connected to disconnected.

re #542