chore(graphql-request): bump graphql-request packages v7.0.1
the package nestjs-graphql-request use the graphql-request and there is a new version (v7.0.1)
available. Because of this new major version (v6 -> v7) I also bump the nestjs-graphql-request to
v1.0.0

BREAKING CHANGE: the 2 main breaking changes are the end of the support for CommonJS + remove
cross-fetch polyfill. The Breaking changes list:
https://github.com/jasonkuhrt/graphql-request/releases/tag/7.0.0

closes #738